requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
pydantic
jose
python-dotenv